Connect message only shows once.
When a players connection attempt is bounced because of the newbie/non-newbie ratio,
the message "[ADMINS]Stripped player bla(bla) connected" will no longer be displayed.
It only gets displayed on a succesful attempt when the player is not bounced.

Every bounced connection attempt is however still printed to the server log.

--Fix by LazyBigCat